rename ViaVai branding to Longlink in web
