+++ web/src/components/longlink/Dialog.tsx
@@ -6,7 +6,6 @@ import {
     DialogFooter,
 } from '@/components/ui/dialog';
 
-
 type DialogProps = {
     trigger: ReactElement;
     confirm?: string;
@@ -16,20 +15,29 @@ type DialogProps = {
     onOpenChange?: (open: boolean) => void;
 };
 
-
 /* 
     A Layout 
 
 */
-export function Dialog({ confirm = 'Confirm', cancel = 'Cancel', children }: DialogProps) {
+export function Dialog({
+    confirm = 'Confirm',
+    cancel = 'Cancel',
+    children,
+}: DialogProps) {
     return (
         <DialogContent className="min-w-[40rem] max-w-5xl">
-            <div className="max-h-[70vh] overflow-y-auto p-1">
-                {children}
-            </div>
+            <div className="max-h-[70vh] overflow-y-auto p-1">{children}</div>
             <DialogFooter>
-                <DialogClose render={<Button variant="outline" className="cursor-pointer" />}>{cancel}</DialogClose>
-                <DialogClose render={<Button className="cursor-pointer" />}>{confirm}</DialogClose>
+                <DialogClose
+                    render={
+                        <Button variant="outline" className="cursor-pointer" />
+                    }
+                >
+                    {cancel}
+                </DialogClose>
+                <DialogClose render={<Button className="cursor-pointer" />}>
+                    {confirm}
+                </DialogClose>
             </DialogFooter>
         </DialogContent>
     );
