+++ web/src/components/longlink/Separator.tsx
@@ -1,7 +1,5 @@
 import { Separator as UISeparator } from '@/components/ui/separator';
 
-
-
 /* 
     A simple seprarator that add a horizontal line between two sections.
 */
